@@ -1,14 +1,14 @@
-import React, { FC, ReactElement, StrictMode, useEffect } from 'react';
-import ReactDOM from 'react-dom';
 import type { Instrument, InstrumentList, Listener } from '@finos/fdc3';
+import React, { FC, ReactElement, StrictMode, useEffect } from 'react';
 
-import { Provider } from '@aiera/client-sdk/components/Provider';
-import { useMessageListener } from '@aiera/client-sdk/lib/msg';
-import { useClient } from '@aiera/client-sdk/api/client';
-import { Auth } from '@aiera/client-sdk/modules/Auth';
 import { defaultTokenAuthConfig } from '@aiera/client-sdk/api/auth';
-import { NewsList } from '@aiera/client-sdk/modules/NewsList';
+import { useClient } from '@aiera/client-sdk/api/client';
+import { Provider } from '@aiera/client-sdk/components/Provider';
 import '@aiera/client-sdk/css/styles.css';
+import { useMessageListener } from '@aiera/client-sdk/lib/msg';
+import { Auth } from '@aiera/client-sdk/modules/Auth';
+import { NewsList } from '@aiera/client-sdk/modules/NewsList';
+import { createRoot } from 'react-dom/client';
 
 const useMessageBus = () => {
     const bus = useMessageListener(
@@ -86,4 +86,6 @@ const App: FC = (): ReactElement => {
     );
 };
 
-ReactDOM.render(<App />, document.getElementById('root'));
+const container = document.getElementById('root');
+const root = createRoot(container!);
+root.render(<App />);
