@@ -1,6 +1,10 @@
-import React, { FC } from 'react';
 import { renderHook } from '@testing-library/react-hooks';
-import { useAuthContext, AuthProvider } from '.';
+import React, { ReactNode } from 'react';
+import { AuthProvider, useAuthContext } from '.';
+
+interface ComponentProps {
+    children: ReactNode;
+}
 
 describe('realtime', () => {
     test('returns undefined with no provider', () => {
@@ -11,7 +15,7 @@ describe('realtime', () => {
     test('returns logout funciton when provided', () => {
         const logout = jest.fn();
 
-        const Wrapper: FC = ({ children }) => {
+        const Wrapper = ({ children }: ComponentProps) => {
             return <AuthProvider logout={logout}>{children}</AuthProvider>;
         };
         const { result } = renderHook(() => useAuthContext(), { wrapper: Wrapper });
